requirements.txt
# Samsung Product Launch Planner - Core Requirements
# Essential dependencies only

# Web Framework
streamlit>=1.28.0

# Data Processing & Analysis
pandas>=1.5.0
numpy>=1.24.0

# Machine Learning & Clustering (Optional)
scikit-learn>=1.3.0

# Visualization
plotly>=5.15.0

# HTTP Requests & APIs
requests>=2.31.0

# Date & Time
python-dateutil>=2.8.2

# Environment Variables
python-dotenv>=1.0.0

# Real-time APIs
pytrends>=4.9.2

# Updated urllib3 for compatibility with pytrends
urllib3>=1.26.12,<2.0

# Optional: Text Processing (will work without)
# textblob>=0.17.1

# Optional: Enhanced visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
Flask>=2.2.5